backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# AI and NLP
anthropic==0.18.1
openai==1.12.0
transformers==4.37.2
torch==2.2.0
scikit-learn==1.4.0
nltk==3.8.1

# Web Scraping
beautifulsoup4==4.12.3
requests==2.31.0
scrapy==2.11.0
newspaper3k==0.2.8
feedparser==6.0.11
lxml==5.1.0

# Task Queue
celery==5.3.6
redis==5.0.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# HTTP Clients
httpx==0.26.0

# Environment
python-dotenv==1.0.1

# CORS
fastapi-cors==0.0.6

# Sentiment Analysis
textblob==0.18.0
vaderSentiment==3.3.2

# Caching
cachetools==5.3.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
